refactor: address code-quality review on the streaming formatter
- Split the 1200-line module into streaming/{mod,printer}.rs: mod.rs holds the
  lexer, the grammar validator, and the entry point; printer.rs holds the emit
  engine. Shared token types are pub(crate).
- Extract object_renders_multiline() (was duplicated between is_breaker and the
  array arm of structurally_multiline) and emit_own_line_comment() (was repeated
  at four comment-placement sites).
- Drop dead code: identity wrapper src_start, unused params, and a hand-rolled
  trim_ascii_end now that std's [u8]::trim_ascii_end is stable.
- Render the flat form once in the object path instead of twice.
- Fix a stale module doc comment (widths are unicode display width, not chars).

No behavior change: full spec suite + error-diagnostic tests still pass.
